ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
pgvector==0.2.3
requests==2.31.0
pydantic==2.5.2
